requirements.txt
requirements: 
1.biopython
2.matplotlib